feat(concurrency): buffered channel holds tokens (#7)
This change introduces a channel to hold tokens created by the lexer and
consumed by the parser.  Upon creation of a new lexer, a go routine is
invoked that begins creating tokens and placing them in the token
buffer up to the specified capacity (default is 10).  Calls to NextToken
return elements from the channel. Upon reaching the end of input, the
channel is closed after writing the EOF token.